fix: use query for meta (#2)
(cherry picked from commit 3bbf41d46771ed98725cd020d7501c0375be8c74)
